feat: 챗봇 뷰 분리 작업
